camera/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.8
numpy==1.26.3
opencv-python==4.10.0.84
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==73.0.1
urllib3==2.2.2
